refactor: refactored cargo workspace to use cargo features feature
