+++ crates/necronux_cli_app/src/commands/lvl_0/necronux.rs
