+++ crates/necronux_cli_app/src/commands/lvl_1/infra.rs
