+++ crates/necronux_cli_app/src/commands/lvl_1/mod.rs
