+++ crates/necronux_cli_app/src/commands/mod.rs
