+++ crates/necronux_cli_app/src/handlers/infra/infra_root.rs
