+++ crates/necronux_cli_app/src/handlers/mod.rs
