+++ crates/necronux_cli_app/src/lib.rs
@@ -4,11 +4,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // ==----------------------------------------------------------------== //
 
-#![deny(clippy::unwrap_used)]
-#![deny(clippy::expect_used)]
-#![deny(clippy::panic)]
-#![deny(unused_must_use)]
-
 mod commands;
 mod flow_controller;
 mod handlers;
@@ -16,7 +11,7 @@ mod logger;
 
 use color_eyre::eyre::Result;
 
-fn main() -> Result<()> {
+pub fn init() -> Result<()> {
     // Initializes the flow controller.
     flow_controller::init_flow_controller()?;
