+++ crates/necronux_cli_app/src/logger.rs
