+++ crates/necronux_internal/src/lib.rs
@@ -4,4 +4,8 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // ==-----------------------------------------------------------== //
 
-pub mod bundle;
+#[cfg(feature = "necronux_cli_app")]
+pub use necronux_cli_app as cli_app;
+#[cfg(feature = "necronux_core")]
+pub use necronux_core as core;
+pub use necronux_utils as utils;
