+++ crates/necronux_utils/src/error_reporter.rs
